pyproject.toml
[project]
name = "pizza-ordering-agent"
version = "0.1.0"
description = "Voice AI agent that calls pizza restaurants and places orders autonomously"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.3",
    "fastapi>=0.127.1",
    "pipecat-ai[deepgram,elevenlabs,google,openai,silero,websocket]>=0.0.102",
    "plivo>=4.59.5",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "uvicorn[standard]>=0.41.0",
]
